fix: renamed SettingsPage and updated import path to follow good practice.
